Add condition for opkg
This PR add condition to avoid issue when opkg is not installed on Creality K1.

+++ install.sh
@@ -185,7 +185,12 @@ ensure_py_venv()
     if [[ $IS_K1_OS -eq 1 ]]
     then
         # The K1 requires we setup the virtualenv like this.
-        virtualenv -p /opt/bin/python3 --system-site-packages "${OCTOAPP_ENV}"
+        if [[ -f /opt/bin/python3 ]]
+        then
+            virtualenv -p /opt/bin/python3 --system-site-packages "${OCTOAPP_ENV}"
+        else
+            python3 /usr/lib/python3.8/site-packages/virtualenv.py -p /usr/bin/python3 --system-site-packages "${OCTOAPP_ENV}"
+        fi
     else
         # Everything else can use this more modern style command.
         virtualenv -p /usr/bin/python3 --system-site-packages "${OCTOAPP_ENV}"
@@ -205,7 +210,10 @@ install_or_update_system_dependencies()
         # the user might install opkg via the 3rd party moonraker installer script.
         # But in general, PY will already be installed, so there's no need to try.
         # On the K1, the only we thing we ensure is that virtualenv is installed via pip.
-        opkg install ${CREALITY_DEP_LIST}
+        if [[ -f /opt/bin/opkg ]]
+        then
+            opkg install ${CREALITY_DEP_LIST}
+        fi
         pip3 install --trusted-host pypi.python.org --trusted-host pypi.org --trusted-host=files.pythonhosted.org --no-cache-dir virtualenv
     elif [[ $IS_SONIC_PAD_OS -eq 1 ]]
     then
